First Run  added Application for HiltViewModel, added module to app module, missing build call on database, added NavHost to Main Activity, modified NoteDetail and NoteScreen,created NoteDetailViewModel, added a new module for local source, created screen routes.

+++ feature/note/src/main/java/com/clisby/shawn/note/presentation/composables/screen/NoteListScreen.kt
@@ -21,7 +21,7 @@ import com.clisby.shawn.note.presentation.model.NoteUiListState
 fun NoteListScreen(
     state: NoteUiListState,
     onAddNote: () -> Unit,
-    onNoteSelected: () -> Unit
+    onNoteSelected: (Int) -> Unit
 ) {
     Scaffold(
         content = {
